verify: relocate verifier to a dedicated module; export receipt.ParsePublicKey
Addresses two code-review items:

#6 (altitude) — move the production browser-verifier source out of the
cross-sdk-tests module (which is linguist-vendored test scaffolding) into a
dedicated obsigna.dev/web-verifier module:

- web-verifier/verifier  — the shared Run core (delegates all crypto to
  obsigna.dev/sdk/go/receipt; reimplements nothing).
- web-verifier/cmd/wasm      (js/wasm)  — browser entry.
- web-verifier/cmd/wasm-cli  (wasip1)   — the gate's WASI twin.

The CLI<->WASM conformance gate stays in cross-sdk-tests (a test, the right
home) and now imports UP into web-verifier, the correct dependency direction.
Wired via go.work; the site build script, the site-obsigna deploy paths, and
the cross-sdk-tests gate paths are updated, and a per-module web-verifier CI
(vet/build + both WASM targets) is added to match the repo convention.

#7 — export receipt.ParsePublicKey and collapse the three hand-rolled
PEM->SPKI->Ed25519 parsers into it: the SDK's was unexported, a cross-sdk test
helper duplicated it, and the verifier hand-rolled a third copy. The verifier
now calls receipt.ParsePublicKey for boundary validation (dropping its
crypto/ed25519, crypto/x509, encoding/pem imports), and the v030/v040/v050
tests use it too.

Verified: gate green (24 cases, builds web-verifier/cmd/wasm-cli via wazero);
sdk/go vet+build+test green; web-verifier vet+build green; the rebuilt browser
artifact verifies single/chain → qualified, tampered/empty-array/bad-key →
fail/error end-to-end under Node.
