+++ .github/workflows/web-verifier.yml
@@ -1,0 +1,44 @@
+name: "CI: web-verifier"
+
+on:
+  push:
+    branches: [main]
+    paths:
+      - "web-verifier/**"
+      - "sdk/go/**"
+  pull_request:
+    branches: [main]
+    paths:
+      - "web-verifier/**"
+      - "sdk/go/**"
+
+permissions:
+  contents: read
+
+jobs:
+  build:
+    runs-on: ubuntu-latest
+    defaults:
+      run:
+        working-directory: web-verifier
+    steps:
+      - uses: actions/checkout@9c091bb21b7c1c1d1991bb908d89e4e9dddfe3e0 # v7.0.0
+
+      - uses: actions/setup-go@4a3601121dd01d1626a1e23e37211e3254c1c06c # v6
+        with:
+          go-version: "1.26"
+
+      - name: Vet
+        run: go vet ./...
+
+      - name: Build
+        run: go build ./...
+
+      # The deployed artifacts. The behavioural CLI<->WASM identity check lives
+      # in cross-sdk-tests (verify_wasm_gate_test.go); here we only confirm both
+      # WASM targets still compile from the shared verifier core.
+      - name: Build browser WASM (js)
+        run: GOOS=js GOARCH=wasm go build ./cmd/wasm
+
+      - name: Build gate WASM (wasip1)
+        run: GOOS=wasip1 GOARCH=wasm go build ./cmd/wasm-cli
