+++ cross-sdk-tests/v050_vectors_test.go
@@ -85,7 +85,7 @@ func TestV050VectorsValidateAgainstSchema(t *testing.T) {
 func TestV050ReceiptHashAndSignature(t *testing.T) {
 	f := loadV050(t)
 
-	pub, err := parseEd25519PublicPEMTest(f.Keys.PublicKey)
+	pub, err := receipt.ParsePublicKey(f.Keys.PublicKey)
 	if err != nil {
 		t.Fatalf("parse public key: %v", err)
 	}
