+++ web-verifier/cmd/wasm/main.go
@@ -1,7 +1,7 @@
 //go:build js && wasm
 
-// Command verify-wasm is the browser entry point for obsigna.dev/verify. It is a
-// thin syscall/js shim over verifier.Run: the page calls the global
+// Command wasm is the browser entry point for obsigna.dev/verify. It is a thin
+// syscall/js shim over verifier.Run: the page calls the global
 // obsignaVerify(requestJSON) and gets back the result JSON. All verification
 // happens in verifier.Run (which delegates to the obsigna.dev/sdk/go/receipt
 // core); this file adds no cryptographic logic.
@@ -10,7 +10,7 @@ package main
 import (
 	"syscall/js"
 
-	"obsigna.dev/cross-sdk-tests/verifier"
+	"obsigna.dev/web-verifier/verifier"
 )
 
 func main() {
