+++ web-verifier/go.mod
@@ -1,0 +1,6 @@
+module obsigna.dev/web-verifier
+
+go 1.26.1
+
+// Depends on obsigna.dev/sdk/go, wired in locally via the repo-root go.work
+// workspace (no require/go.sum needed — never published, workspace-only).
